requirements.txt
# FastAPI
fastapi
uvicorn[standard]
python-multipart
aiofiles

# DeepLearning
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
scipy>=1.4.1
torch>=1.7.0,!=1.12.0
torchvision>=0.8.1,!=0.13.0
thop
ffmpeg-python
pandas
seaborn
tqdm